% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_addcalendarevent.R
\name{suso_gql_addinterviewcalendarevent}
\alias{suso_gql_addinterviewcalendarevent}
\title{Add a calendar event to an Interview}
\usage{
suso_gql_addinterviewcalendarevent(
  endpoint = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  token = NULL,
  interviewId = NULL,
  comment = NULL,
  newStart = NULL,
  startTimezone = "UTC"
)
}
\arguments{
\item{endpoint}{GraphQL endpoint of your server}

\item{workspace}{Server Workspace, if NULL uses default}

\item{user}{your API username}

\item{password}{your API user password}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{interviewId}{the interviewId}

\item{comment}{a comment string}

\item{newStart}{new start date, format must be: \code{2024-01-16 01:41:14}}

\item{startTimezone}{time zone of the tablet device, use \code{\link[base]{OlsonNames}}}
}
\description{
Add a calendar event to an Interview
}
