% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_updatecalendarevent.R
\name{suso_gql_deletecalendarevent}
\alias{suso_gql_deletecalendarevent}
\title{Delete a calendar event}
\usage{
suso_gql_deletecalendarevent(
  endpoint = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  token = NULL,
  publicKey = NULL
)
}
\arguments{
\item{endpoint}{GraphQL endpoint of your server.}

\item{workspace}{Server Workspace, if NULL uses default.}

\item{user}{your API username.}

\item{password}{your API user password.}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored.}

\item{publicKey}{the publicKey of the calendar event to be deleted.}
}
\description{
Delete a calendar event
}
