% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_deleteuserfrommap.R
\name{suso_gql_deleteuserfrommap}
\alias{suso_gql_deleteuserfrommap}
\title{Delete user from map}
\usage{
suso_gql_deleteuserfrommap(
  endpoint = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  fileName = NULL,
  userName = NULL
)
}
\arguments{
\item{endpoint}{GraphQL endpoint of your server}

\item{workspace}{Server Workspace, if NULL uses default}

\item{user}{your API username}

\item{password}{your API user password}

\item{fileName}{the name of the map file on the server}

\item{userName}{the name of the interviewer to whom the map will be assigned to}
}
\description{
Allows the user to delete an interviewer from a map to be used in CAPI data collection.
}
