% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_credentials.R
\name{suso_gql_pwcheck}
\alias{suso_gql_pwcheck}
\title{Utility function to check if credentials are correct}
\usage{
suso_gql_pwcheck(
  endpoint = "https://demo.mysurvey.solutions/graphql",
  user = "someuser",
  password = "andhispassword",
  workspace = "primary",
  token = NULL
)
}
\arguments{
\item{endpoint}{Survey Solutions GraphQl}

\item{user}{API user}

\item{password}{API password}

\item{workspace}{endpoint workspace Name, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions endpoint token is provided \emph{user} and \emph{password} will be ignored}
}
\value{
200 code if correct, 400 if incorrect.
}
\description{
This function returns a 200 status code if credentials are correct and a 400 code otherwise.
}
\details{
If the app runs interactively, status is printed to the console, if it runs in a shiny app, a status
notification will be shown, if option \option{suso.useshiny} is \code{TRUE}.
}
