% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_questionnaires.R
\name{suso_gql_questionnaires}
\alias{suso_gql_questionnaires}
\title{Get all Questionnaires}
\usage{
suso_gql_questionnaires(
  endpoint = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  id = NULL,
  version = NULL
)
}
\arguments{
\item{endpoint}{GraphQL endpoint of your server}

\item{workspace}{Server Workspace, if NULL uses default}

\item{user}{your API username}

\item{password}{API password}

\item{id}{Questionnaire ID}

\item{version}{Questionnaire version}
}
\description{
Allows the user to retrieve filtered or unfiltered map data.
}
