% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_questions.R
\name{suso_gql_questions}
\alias{suso_gql_questions}
\title{Get all Questions in a Questionnaires}
\usage{
suso_gql_questions(
  endpoint = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  id = NULL,
  version = NULL,
  variable = NULL,
  scope = NULL,
  identifying = NULL
)
}
\arguments{
\item{endpoint}{GraphQL endpoint of your server}

\item{workspace}{Server Workspace, if NULL uses default}

\item{user}{your API username}

\item{password}{API password}

\item{id}{Questionnaire ID}

\item{version}{Questionnaire version}

\item{variable}{Get questions for a specific variable}

\item{scope}{Get questions for a specific scope}

\item{identifying}{If TRUE only identifying questions are exported}
}
\description{
Allows the user to retrieve all questions in a questionniare.
}
