% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_transformers_selector.R
\docType{data}
\name{susoop_str}
\alias{susoop_str}
\title{List object for Survey Solutions GraphQl character operator selection}
\format{
An object of class \code{list} of length 10.
}
\usage{
susoop_str
}
\value{
A named list with the operator and the value to be passed on as input to
the filter.
}
\description{
A list of the available transformers
}
\details{
Allows the user to select the operator for the required
filter.
}
\examples{

# equal to 3
susoop_str$contains("area10")

# not equal to 3
susoop_str$startsWith("area")

}
\keyword{datasets}
