% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_transformers.R
\name{transformers}
\alias{transformers}
\alias{eq}
\alias{neq}
\alias{inbetw}
\alias{ninbetw}
\alias{gt}
\alias{ngt}
\alias{gte}
\alias{ngte}
\alias{lt}
\alias{nlt}
\alias{lte}
\alias{nlte}
\title{Utility Functions for operator selection}
\usage{
eq(value_set)

neq(value_set)

inbetw(value_set)

ninbetw(value_set)

gt(value_set)

ngt(value_set)

gte(value_set)

ngte(value_set)

lt(value_set)

nlt(value_set)

lte(value_set)

nlte(value_set)
}
\arguments{
\item{value_set}{the parameter set for the operator}
}
\description{
Can be used in filters ("where") for operator selection. If none is selected, operator always defaults to `eq()`. The functions bellow are
valid for the corresponding inputs ComparableInt64OperationFilterInput and ComparableNullableOfInt32OperationFilterInput.
}
\section{Functions}{
\itemize{
\item \code{neq()}: smaller

\item \code{inbetw()}: inbetween (requires numeric vector with upper and lower bound, i.e. c(1, 5))

\item \code{ninbetw()}: not inbetween (requires numeric vector with upper and lower bound, i.e. c(1,5))

\item \code{gt()}: greater than

\item \code{ngt()}: not greater than

\item \code{gte()}: greater than or equal

\item \code{ngte()}: not greater than or equal

\item \code{lt()}: lower than

\item \code{nlt()}: not lower than

\item \code{lte()}: lower than or equal

\item \code{nlte()}: not lower than or equal

}}
\examples{

\dontrun{
suso_gql_interviews(endpoint = [srv], user = [user],
password = [pass], workspace = "test", errorsCount = eq(0))


suso_gql_interviews(endpoint = [srv], user = [user],
password = [pass], workspace = "test", errorsCount = gt(0))


suso_gql_interviews(endpoint = [srv], user = [user],
password = [pass], workspace = "test", errorsCount = inbetw(c(0, 5))
}

}
