% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_assignments.R
\name{suso_gql_assignments}
\alias{suso_gql_assignments}
\title{Get all assignments}
\usage{
suso_gql_assignments(
  endpoint = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  archived = FALSE,
  id = NULL,
  questionnaireId = NULL,
  version = NULL,
  responsibleId = NULL,
  webMode = FALSE,
  take = NULL,
  skip = NULL
)
}
\arguments{
\item{endpoint}{GraphQL endpoint of your server}

\item{workspace}{Server Workspace, if NULL uses default}

\item{user}{your API username}

\item{password}{API password}

\item{archived}{Boolean if the assignment is archived or not}

\item{id}{Assignment ID}

\item{questionnaireId}{Questionnaire ID}

\item{version}{Questionnaire version}

\item{responsibleId}{ID of the person currently responsible for the assignment}

\item{webMode}{Boolean for webmode}

\item{take}{take the specified integer numeber of maps}

\item{skip}{skip the first integer number of maps}
}
\description{
Allows the user to retrieve filtered or unfiltered assignment data.
}
