% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_updatecalendarevent.R
\name{suso_gql_deletecalendarevent}
\alias{suso_gql_deletecalendarevent}
\title{Delete a calendar event}
\usage{
suso_gql_deletecalendarevent(
  endpoint = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  token = NULL,
  publicKey = NULL
)
}
\arguments{
\item{endpoint}{GraphQL endpoint of your server.}

\item{workspace}{Server Workspace, if NULL uses default.}

\item{user}{your API username.}

\item{password}{your API user password.}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored.}

\item{publicKey}{the publicKey of the calendar event to be deleted.}
}
\value{
if successfull, returns a list with the executed mutation
}
\description{
Delete a calendar event
}
\examples{
\dontshow{if (suso_gql_pwcheck()==200) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Requires Survey Solutions Server and API credentials

# Delete a calendar event for a specific event
# in timezone Chile/Continental

suso_gql_deletecalendarevent(endpoint = ep, user = usr,
password = pass, workspace = ws,
publicKey = "ea8b49c0-0d16-4fec-9029-72278e94a233")
\dontshow{\}) # examplesIf}
}
