% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_deleteMap.R
\name{suso_gql_deletemap}
\alias{suso_gql_deletemap}
\title{Delete map}
\usage{
suso_gql_deletemap(
  endpoint = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  token = NULL,
  fileName = NULL
)
}
\arguments{
\item{endpoint}{GraphQL endpoint of your server}

\item{workspace}{Server Workspace, if NULL uses default}

\item{user}{your API username}

\item{password}{your API user password}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{fileName}{the name of the map file on the server}
}
\value{
if successfull, returns a list with the executed mutation
}
\description{
Allows the user to delete a map.
}
\examples{
\dontshow{if (suso_gql_pwcheck()==200) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Requires Survey Solutions Server and API credentials

# Delete map seg_168_ALL.tif

suso_gql_deletemap(endpoint = ep, user = usr,
password = pass, workspace = ws,
fileName = "seg_168_ALL.tif")

\dontshow{\}) # examplesIf}
}
