% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_interviews.R
\name{suso_gql_interviews}
\alias{suso_gql_interviews}
\title{Get all interviews}
\usage{
suso_gql_interviews(
  endpoint = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  assignmentId = NULL,
  clientKey = NULL,
  createdDate = NULL,
  errorsCount = NULL,
  identifyingData = NULL,
  interviewMode = NULL,
  notAnsweredCount = NULL,
  questionnaireId = NULL,
  questionnaireVersion = NULL,
  questionnaireVariable = NULL,
  responsibleName = NULL,
  responsibleRole = NULL,
  status = NULL,
  supervisorName = NULL
)
}
\arguments{
\item{endpoint}{GraphQL endpoint of your server}

\item{workspace}{Server Workspace, if NULL uses default}

\item{user}{your API username}

\item{password}{API password}

\item{assignmentId}{Assignment ID}

\item{clientKey}{Interview key}

\item{createdDate}{Creation data of the interview}

\item{errorsCount}{number of errors}

\item{identifyingData}{Pre-loaded identifying data}

\item{interviewMode}{Interview mode (CAWI or CAPI)}

\item{notAnsweredCount}{number of unanswered questions}

\item{questionnaireId}{the questionnnaire id}

\item{questionnaireVersion}{the questionnaire version}

\item{questionnaireVariable}{the variable for the questionnaire}

\item{responsibleName}{Name of the person responsible}

\item{responsibleRole}{Role of the person responsible}

\item{status}{of the interview}

\item{supervisorName}{Name of the supervisor of the responsible user}
}
\description{
Allows the user to retrieve filtered or unfiltered interview data.
}
