% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_maps.R
\name{suso_gql_maps}
\alias{suso_gql_maps}
\title{Get all maps}
\usage{
suso_gql_maps(
  endpoint = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  fileName = NULL,
  importDateUtc = NULL,
  size = NULL,
  users = NULL
)
}
\arguments{
\item{endpoint}{GraphQL endpoint of your server}

\item{workspace}{Server Workspace, if NULL uses default}

\item{user}{your API username}

\item{password}{API password}

\item{fileName}{name of the map on the server}

\item{importDateUtc}{Import date}

\item{size}{Size of the map}

\item{users}{Users to whom the maps are assigned}
}
\description{
Allows the user to retrieve filtered or unfiltered map data.
}
