% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_questionnaires.R
\name{suso_gql_questionnaires}
\alias{suso_gql_questionnaires}
\title{Get all Questionnaires}
\usage{
suso_gql_questionnaires(
  endpoint = NULL,
  workspace = NULL,
  user = NULL,
  password = NULL,
  token = NULL,
  id = NULL,
  version = NULL,
  take = NULL,
  skip = NULL
)
}
\arguments{
\item{endpoint}{GraphQL endpoint of your server}

\item{workspace}{Server Workspace, if NULL uses default}

\item{user}{your API username}

\item{password}{API password}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{id}{Questionnaire ID}

\item{version}{Questionnaire version}

\item{take}{take the specified integer numeber of questionnaires}

\item{skip}{skip the first integer number of questionnaires}
}
\description{
Allows the user to retrieve filtered or unfiltered map data.
}
