% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_gql_uploadMap.R
\name{suso_gql_uploadmap}
\alias{suso_gql_uploadmap}
\title{Upload map to server}
\usage{
suso_gql_uploadmap(
  endpoint = NULL,
  workspace = NULL,
  user = NULL,
  token = NULL,
  password = NULL,
  path_to_zip = NULL,
  usePar = FALSE,
  n_par = 10
)
}
\arguments{
\item{endpoint}{GraphQL endpoint of your server.}

\item{workspace}{Server Workspace, if NULL uses default.}

\item{user}{your API username.}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored.}

\item{password}{API password.}

\item{path_to_zip}{path to a single zip file or path to a directory of zip files if \code{usePar=TRUE}.}

\item{usePar}{if TRUE the requests will be performed in parallel EXPERIMENTAL!}

\item{n_par}{number of parallel requests, required if \code{usePar = TRUE} EXPERIMENTAL!}
}
\value{
list with details on successfully processed maps.
}
\description{
Allows the user to upload 1 or several zip files with base maps and shapefiles to the server.
}
